backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pymysql
cryptography
pydantic[email]
python-dotenv
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
python-jose[cryptography]
python-multipart
email-validator